+++ game_patch/graphics/d3d11/gr_d3d11_texture.h
@@ -48,6 +48,7 @@ namespace df::gr::d3d11
         void finish_render_target(int bm_handle);
         void save_cache();
         void flush_cache(bool force);
+        void flush_non_user_cache();
         void add_ref(int bm_handle);
         void remove_ref(int bm_handle);
         void mark_dirty(int bm_handle);
@@ -149,6 +150,7 @@ namespace df::gr::d3d11
         }
 
         Texture create_texture(int bm_handle, rf::bm::Format fmt, int w, int h, rf::ubyte* bits, rf::ubyte* pal, int mip_levels, bool staging, int src_w = 0, int src_h = 0);
+        Texture create_texture_auto_mips(int bm_handle, rf::bm::Format fmt, int w, int h, rf::ubyte* bits, rf::ubyte* pal);
         Texture create_render_target(int bm_handle, int w, int h);
         Texture load_texture(int bm_handle, bool staging);
         std::pair<DXGI_FORMAT, rf::bm::Format> determine_supported_texture_format(rf::bm::Format fmt);