+++ docs/CHANGELOG.md
@@ -104,6 +104,7 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Add `Gas_Region_State`, `Modify_Gas_Region`, `Resize_Gas_Region` events
 - Add `require_d3d11` to dedicated server config
 - Add `$Stock Alpha Test` to `af_level_quirks.tbl` and use to restore stock alpha test only on known affected levels
+- Serve detailed player and server info in `players_request` packet response (for online FF RFSB)
 - Support 8-bit greyscale TGA textures (types 3 and 11) in editor and game
 - Add experimental D3D11 performance settings `D3D11_LowFrameLatency`, `D3D11_AllowTearing`, and `D3D11_SkipGammaPass`
 
@@ -191,6 +192,7 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Fix skybox rendering issues with Direct3D 11 renderer on community level `DM-Blunderscannon~~.rfl`
 - Fix clutter death achievement progress and `AF_When_Dead` not working on all clutter death paths
 - Fix level editor crash when manipulating decals in levels with > 127 decals
+- Fix `r_picmip` not properly applying to all mesh textures
 
 [@is-this-c](https://github.com/is-this-c)
 - Fix parse of `flag_return_time` to be as a float instead of an integer